requirements.txt
# local package
-e .

# external requirements cookie cutter datascience
click
Sphinx
coverage
awscli==1.27.105
flake8
python-dotenv>=0.5.1

# external requirements 
langchain
openai
chromadb
beautifulsoup4
tqdm